requirements.txt
django-qrcode
Django==1.11.6
django-extensions==1.9.1
django-localflavor==1.5.2
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
djdt_flamegraph
django-silk
Pygments>=2.0
python-dateutil>=2.4
requests>=2.10
sqlparse>=0.1.19
Jinja2>=2.8
autopep8>=1.2.1
pytz>2014.2
mock>=1.0.1
Pillow>=3.2
django>=1.8
freezegun>=0.3
factory-boy>=2.8.1
gprof2dot>=2016.10.13,<2017.09.19